# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/room_chat": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roomChat"
                ],
                "summary": "get room chat by session ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "filter session status",
                        "name": "session_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RoomChatDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.HttpErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpErrorModel"
                        }
                    }
                }
            }
        },
        "/room_chats": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roomChat"
                ],
                "summary": "get room chat list by participant ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "participant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "filter session status",
                        "name": "session_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RoomChatDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.HttpErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ChatDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "messageType": {
                    "type": "string"
                },
                "readAt": {
                    "type": "string"
                },
                "replyforChatId": {
                    "type": "string"
                },
                "senderId": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "dto.RoomChatDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "participantGroup": {
                    "type": "string"
                },
                "sessionChats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SessionChatDto"
                    }
                }
            }
        },
        "dto.SessionChatDto": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatDto"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "endedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomChatId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "http.HttpErrorModel": {
            "type": "object",
            "properties": {
                "code": {},
                "httpStatus": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}